% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cells_to_unlock.R
\name{find_cells_to_unlock}
\alias{find_cells_to_unlock}
\title{Find indices of rows and columns of data to be unlocked}
\usage{
find_cells_to_unlock(data, ...)
}
\arguments{
\item{data}{data containing the cells to be unlock}

\item{...}{column names where some cells should be unlocked}
}
\value{
tibble containing all rows and columns information about
which cell should be locked or unlocked
}
\description{
Find indices of rows and columns of data to be unlocked
}
\examples{
#' @examples
#' \dontrun{
#' data <- tibble::tribble(
#'   ~a, ~year, ~f, ~d, ~b, ~c,
#'   NA, "2019", "g", NA, "f", NA,
#'   "e", NA, "g", "e", NA, NA,
#'   NA, "2021", "g", "e", "f", NA,
#' )
#' find_cells_to_unlock(data = data, "a", "year", "b", "c")
#' }
}
