% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.demo-package.R
\docType{package}
\name{openxlsx.demo-package}
\alias{openxlsx.demo}
\alias{openxlsx.demo-package}
\title{openxlsx.demo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Layal Christine Lettry \email{layal@cynkra.com}

}
\keyword{internal}
