% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_penguins_mod.R
\name{prepare_penguins_mod}
\alias{prepare_penguins_mod}
\title{Prepare the data penguins}
\usage{
prepare_penguins_mod(data, data_raw)
}
\arguments{
\item{data}{tibble data_penguins}

\item{data_raw}{tibble data_penguins_raw}
}
\value{
tibble
}
\description{
Prepare the data penguins
}
\examples{
#' @examples
#' \dontrun{
#' prepare_penguins_mod(data_penguins = palmerpenguins::penguins)
#' }
}
