% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hyperlink.R
\name{write_hyperlink}
\alias{write_hyperlink}
\title{Write internal hyperlinks}
\usage{
write_hyperlink(dataset, metadata, excel_sheet, first_row, meta_ws_name, wb)
}
\arguments{
\item{dataset}{dataset containing the any_comment variable to link with the one in
the metadata dataset}

\item{metadata}{metadata dataset where the source variable Comments is stored}

\item{excel_sheet}{name of the excel sheet where the internal link should be
written}

\item{first_row}{first row where the data should be written in the sheet}

\item{meta_ws_name}{name of the sheet containing the metadata dataset}

\item{wb}{name of excel workbook}
}
\value{
invisible
}
\description{
Write internal hyperlinks between the meta_ws_name sheet and another one to
link the Comments variable to the source variable in meta_ws_name.
}
