% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openxlsx.demo-package.R
\docType{package}
\name{openxlsx.demo-package}
\alias{openxlsx.demo}
\alias{openxlsx.demo-package}
\title{openxlsx.demo: openxlsx demonstration}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create an excel file from scratch only using openxlsx and adding some features one could be interested in.
}
\author{
\strong{Maintainer}: Layal Christine Lettry \email{layal@cynkra.com}

}
\keyword{internal}
