% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_penguins.R
\name{write_penguins}
\alias{write_penguins}
\title{Write penguins datasets into an xlsx file using openxlsx.}
\usage{
write_penguins(data_penguins, data_penguins_raw, folder_xlsx_file)
}
\arguments{
\item{data_penguins}{dataset penguins.}

\item{data_penguins_raw}{dataset penguins_raw.}

\item{folder_xlsx_file}{directory of the folder where you want to save the
xlsx file.}
}
\value{
invisible
}
\description{
Write the palmerpenguins::penguins and palmerpenguins::penguins_raw
datasets in an excel file and add some features.
}
\details{
Idea: create an excel file from scratch only using openxlsx and adding some
features one could be interested in.

For our first worksheet, we will use the palmerpenguins::penguins data set
(species, island, bill length,bill depth, flipper length, body mass in grams,
sex, year). Ideally, we want to modify it and add some features.

For the second sheet, we will use the palmerpenguins::penguins_raw data set which
contains all the variables and original names as downloaded (please refer to
https://allisonhorst.github.io/palmerpenguins/ for more details).
}
\examples{
#' @examples
#' \dontrun{
#' write_penguins(
#'   data_penguins = palmerpenguins::penguins,
#'   data_penguins_raw = palmerpenguins::penguins_raw,
#'   folder_xlsx_file = tempdir()
#' )
#' }
}
